% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{format_tel}
\alias{format_tel}
\title{Formatage numéros de téléphone}
\usage{
format_tel(num)
}
\arguments{
\item{num}{Vecteur de numéros de télephone en numérique ou caractère,
le 0 initial est facultatif}
}
\description{
Mise en forme de numéros de téléphones.
}
\details{
Defaut au format français: 06 ## ## ## ##.
Détection des numéros réunionnais, commençant par 0262 ou 0692 et
formaté ainsi: 0262 ## ## ##.
}
\examples{
  format_tel(c('0745623118', 262856231, ' 0692856214'))
}
