% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{indic}
\alias{indic}
\title{Indicateurs évaluation des modèles}
\usage{
indic(obs, sim, indicateur, .poids = NULL)
}
\arguments{
\item{obs}{Vecteur des valeurs observées.}

\item{sim}{Vecteur des valeurs simulées.}

\item{indicateur}{Nom de l'indicateur à calculer. Non sensible à la casse.
\itemize{
\item variables quantitatives: RMSE, RRMSE, MAE, Efficience (eff), Biais, MAPE, RMSPE.
\item variables qualitatives: Specificite (spe), Sensibilite (sen, sens),
Taux_bien_predit (tbp), Youden_index (J)
}}

\item{.poids}{Poids à attribuer aux valeur dans le calcul de l'indicateur.}
}
\description{
Calculs les indicateurs utile à l'évaluation de la performance des modèles.
}
\details{
Indicateurs pour variables Quantitatives
\itemize{
\item RMSE : Root Mean Square Error. L'indicateur le plus courant.
Erreur moyenne en augmentant le poids des grosses erreurs via le carré.
\item RRMSE : RMSE / moyenne des observations. Idem mais transformé en indice
pour pouvoir être comparé à des modèles utilisant des données de moyenne différente.
\item MAE : Mean Absolute Error. Erreur moyenne sans augmenter le poids des
grosses erreurs.
\item Efficience (ou eff) : erreur au carré moyenne / equart à la moyenne au carré.
A quel point le modèle est meilleur que de juste utiliser la moyenne.
\item Biais : moyenne des erreurs. Le modèle dévit-il ?
\item MAPE : Mean Absolute Percentage Error.
Equivalent MAE pour des pourcentages d'erreur.
Utile pour des données ayant une grande amplitude,
typiquement des modèles de croissance.
\item RMSPE : Root Mean Squared Percentage Error (indicateur perso je crois)
Equivalent RMSE pour des pourcentages d'erreur.
Utile pour des données ayant une grande amplitude,
typiquement des modèles de croissance.
}

Indicateurs pour variables Qualitatives
\itemize{
\item Specificite (ou spe) : taux de prévision des négatifs
\item Sensibilite (ou sen, sens) : taux de prévision des positifs
\item Taux_bien_predit (ou tbp) : Classique mais trompeur si les données sont
biaisées en faveur des positifs ou des négatifs
\item Youden_index (ou J) : sensibilité + spécificité - 1
Indice de performance global, insensible au bias des données d'entrainement
}
}
