% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{lire_onglet}
\alias{lire_onglet}
\title{Lire l'onglet d'un classeur Excel}
\usage{
lire_onglet(classeur, onglet, tableau = TRUE)
}
\arguments{
\item{classeur}{Nom du Classeur Excel.}

\item{onglet}{Nom de l'onglet à importer.}

\item{tableau}{Si \code{TRUE}, force la récupération des numériques non arrondi
même lorsque leurs colones contiennent du texte.}
}
\description{
Appels de \code{XLConnect::readWorksheet} pré-parametré
}
