% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{loca_dossier}
\alias{loca_dossier}
\title{Localise un fichier ou un dossier}
\usage{
loca_dossier(Dossiers, marqueur, Stop = "\\\\*")
}
\arguments{
\item{Dossiers}{La direction du dossier dans lequel chercher.}

\item{marqueur}{Regex identifiant le/les fichiers cherchées.}

\item{Stop}{Regex pour identifier les dossiers dans lesquels ne pas chercher.}
}
\value{
vecteur des fichiers identifiés
}
\description{
Fonction recursive :) premettant d'identifier tous les dossiers et/ou fichiers
correspondant au marqueur renseigné.
}
