% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pvalue_signif}
\alias{pvalue_signif}
\title{Significativité d'une pvalue}
\usage{
pvalue_signif(pvalue)
}
\arguments{
\item{pvalue}{valeur unique}
}
\description{
Mise en forme de la significativité d'une pvalue
}
\details{
\itemize{
\item pvalue > 0.1 : ' '
\item 0.1 >= pvalue > 0.05 : .
\item 0.05 >= pvalue > 0.01 : *
\item 0.01 >= pvalue > 0.001 : **
\item 0.001 >= pvalue : ***
}
}
