% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{st_read_ftp}
\alias{st_read_ftp}
\title{Lecture de shapefile depuis un serveur FTP}
\usage{
st_read_ftp(url_dossier, identification = NULL)
}
\arguments{
\item{url_dossier}{URL du dossier contenant le shapefile,
format \verb{ftp://username:password@ftppath} seul ou \verb{ftp://ftppath}
en précisant \code{identification} si nécessaire}

\item{identification}{\code{username:password} du FTP si nécéssaire
uniquement si ils ne sont pas intégrés dans \code{url_dossier}}
}
\value{
objet de classe \code{sf}, voir doc du package \code{sf}
}
\description{
Détection, téléchargement depuis un serveur FTP puis lecture des fichiers shapefile.
}
\details{
La détéction des fichiers se fait avec \code{RCurl::getURL}.
Ils sont téléchargé dans une direction temporaire et
lu avec \code{sf::st_read}
}
