% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{stand_commune}
\alias{stand_commune}
\title{Standardisation noms de communes (La Réunion)}
\usage{
stand_commune(Commune, champs = c("code_insee", "Commune"))
}
\arguments{
\item{Commune}{noms de communes avec typologie chaotique}

\item{champs}{type de résultats attendus code_insee, Commune et/ou Commune_MAJ}
}
\description{
Standardisation noms de communes (La Réunion)
}
