% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajorant.divers-package.R
\docType{package}
\name{ajorant.divers-package}
\alias{ajorant.divers}
\alias{ajorant.divers-package}
\title{ajorant.divers: Divers fonctions utilitaires inclassables}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Amaury Jorant \email{amaury.jorant@protonmail.com}

}
\keyword{internal}
